requirements.txt
numpy==2.1.1
pandas==2.2.3
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
xlrd==2.0.1